validator: move the remaining groups into a standard module
What is left at the source root is every group that only reads and
writes its own schema, so one cluster can serve them all. The umbrella
group comes later, with the cluster it owns.

Move only.
