+++ crates/validator/src/standard/alternator/batch_write_item.rs
