+++ crates/validator/src/standard/mod.rs
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2026-present ScyllaDB
+ * SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.1
+ */
+
+mod alternator;
+mod ann;
+mod cdc;
+mod coexisting_indexes;
+mod crud;
+mod filtering;
+mod fts;
+mod index_create;
+mod index_modify;
+mod index_status;
+mod quantization_and_rescoring;
+mod serde;
+mod similarity_functions;
